+++ packages/components/charts/src/pie-chart.tsx
@@ -94,6 +94,8 @@ export const PieChart = forwardRef<PieChartProps, "div">((props, ref) => {
     endAngle,
     withLabels,
     withLabelLines,
+    labelOffset,
+    isParcent,
     strokeWidth,
     legendProps,
     ...rest
@@ -118,6 +120,9 @@ export const PieChart = forwardRef<PieChartProps, "div">((props, ref) => {
     strokeWidth,
     withLabels,
     withLabelLines,
+    labelOffset,
+    isParcent,
+    valueFormatter,
     styles,
   })
   const { getContainerProps } = useChart({ containerProps })
@@ -160,8 +165,6 @@ export const PieChart = forwardRef<PieChartProps, "div">((props, ref) => {
             <Pie
               {...getPieProps({
                 className: "ui-pie-chart__pie",
-                labelClassName: "ui-pie-chart__label",
-                labelLineClassName: "ui-pie-chart__label-line",
               })}
             >
               {cells}