+++ packages/components/visually-hidden/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @yamada-ui/visually-hidden
 
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies [[`7765580`](https://github.com/yamada-ui/yamada-ui/commit/776558080aa0d07bf497c5b24b244e8e28b6b876)]:
+  - @yamada-ui/core@1.10.0
+
 ## 1.0.5
 
 ### Patch Changes